Update services-behaviors.md mergeConfig reference
In this documenation block, specifically the `Detecting utilized extensions` part, the text refers to `$this->mergeConfig()`. The reference to `$this` is only valid if the extension is done inside another controller, which is not the place where you would extend the controller.

By using the `$controller` reference, it is made sure that the `mergeConfig` method is always available. This is backed up by the answer on #3105.