Added FormWidget Directory structure example
I had a hard time identifying this problem when creating a Form Widget. Once you realize it it's obvious, but it might take some time to get a grasp of why things are not working fully. Adding this block of docs will help those like me who are on a clock trying to figure out these things.