Update belongsToMany with Pivot data note
The current documentation is slightly ambiguous when describing the consequences of enabling `deferredBinding` in the relation behavior config for belongsToMany with Pivot data. This PR updates the note to include more details, and also moves it to the top of the Belongs to many (with pivot data) section so that it is more relevant.