Update backend list filter scope options
This PR updates the definition for the scope option on backend lists. The current definition is incorrect in that it states the filtered values are passed to the specified scope method as the first parameter, when in fact, they are passed as the second parameter. The first parameter actually contains the query.